split out action_addlist.go
