+++ action.go
@@ -1598,91 +1598,3 @@ func (a *ActionGetLinks) Receive(dht *DHT, msg *Message, retries int) (response 
 
 	return
 }
-
-//------------------------------------------------------------
-// ListAdd
-
-type ActionListAdd struct {
-	list PeerList
-}
-
-func NewListAddAction(peerList PeerList) *ActionListAdd {
-	a := ActionListAdd{list: peerList}
-	return &a
-}
-
-func (a *ActionListAdd) Name() string {
-	return "put"
-}
-
-func (a *ActionListAdd) Args() []Arg {
-	return nil
-}
-
-func (a *ActionListAdd) Do(h *Holochain) (response interface{}, err error) {
-	err = NonCallableAction
-	return
-}
-
-var prefix string = "List add request rejected on warrant failure"
-
-func (a *ActionListAdd) Receive(dht *DHT, msg *Message, retries int) (response interface{}, err error) {
-	t := msg.Body.(ListAddReq)
-	a.list.Type = PeerListType(t.ListType)
-	a.list.Records = make([]PeerRecord, 0)
-	var pid peer.ID
-	for _, pStr := range t.Peers {
-		pid, err = peer.IDB58Decode(pStr)
-		if err != nil {
-			return
-		}
-		r := PeerRecord{ID: pid}
-		a.list.Records = append(a.list.Records, r)
-	}
-
-	// validate the warrant sent with the list add request
-	var w Warrant
-	w, err = DecodeWarrant(t.WarrantType, t.Warrant)
-	if err != nil {
-		err = fmt.Errorf("%s: unable to decode warrant (%v)", prefix, err)
-		return
-	}
-
-	err = w.Verify(dht.h)
-	if err != nil {
-		err = fmt.Errorf("%s: %v", prefix, err)
-		return
-	}
-
-	// TODO verify that the warrant, if valid, is sufficient to allow list addition #300
-
-	err = dht.addToList(msg, a.list)
-	if err != nil {
-		return
-	}
-
-	// special case to add blockedlist peers to node cache and delete them from the gossipers list
-	if a.list.Type == BlockedList {
-		for _, node := range a.list.Records {
-			dht.h.node.Block(node.ID)
-			dht.DeleteGossiper(node.ID) // ignore error
-		}
-	}
-	response = DHTChangeOK
-	return
-}
-
-// retryIfHashNotFound checks to see if the hash is found and if not queues the message for retry
-func (dht *DHT) retryIfHashNotFound(hash Hash, msg *Message, retries int) (response interface{}, err error) {
-	err = dht.exists(hash, StatusDefault)
-	if err != nil {
-		if err == ErrHashNotFound {
-			dht.dlog.Logf("don't yet have %s, trying again later", hash)
-			retry := &retry{msg: *msg, retries: retries}
-			dht.retryQueue <- retry
-			response = DHTChangeUnknownHashQueuedForRetry
-			err = nil
-		}
-	}
-	return
-}
